# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AuthEngine",
    "description": "Production-ready IAM system built with FastAPI — supports authentication, multi-tenancy, fine-grained permissions, and token introspection to power multiple apps from a unified identity layer.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/tenants/{tenant_id}/email-config": {
      "get": {
        "tags": ["tenant-email-config"],
        "summary": "Get tenant email configuration",
        "operationId": "get_email_config_api_v1_tenants__tenant_id__email_config_get",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/TenantEmailConfigResponse"
                    },
                    {
                      "$ref": "#/components/schemas/TenantEmailConfigFallbackResponse"
                    }
                  ],
                  "title": "Response Get Email Config Api V1 Tenants  Tenant Id  Email Config Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["tenant-email-config"],
        "summary": "Create tenant email configuration",
        "operationId": "create_email_config_api_v1_tenants__tenant_id__email_config_post",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantEmailConfigCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantEmailConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["tenant-email-config"],
        "summary": "Update tenant email configuration",
        "operationId": "update_email_config_api_v1_tenants__tenant_id__email_config_put",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantEmailConfigUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantEmailConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["tenant-email-config"],
        "summary": "Delete tenant email configuration (falls back to platform default)",
        "operationId": "delete_email_config_api_v1_tenants__tenant_id__email_config_delete",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "204": { "description": "Successful Response" },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/tenants/{tenant_id}/email-config/test": {
      "post": {
        "tags": ["tenant-email-config"],
        "summary": "Send a test email using the tenant's configured provider",
        "operationId": "test_email_config_api_v1_tenants__tenant_id__email_config_test_post",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailConfigTestRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailConfigTestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EmailConfigTestRequest": {
        "properties": {
          "to_email": {
            "type": "string",
            "format": "email",
            "title": "To Email"
          }
        },
        "type": "object",
        "required": ["to_email"],
        "title": "EmailConfigTestRequest",
        "description": "Body for POST /tenants/{tenant_id}/email-config/test"
      },
      "EmailConfigTestResponse": {
        "properties": {
          "success": { "type": "boolean", "title": "Success" },
          "error": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Error"
          }
        },
        "type": "object",
        "required": ["success"],
        "title": "EmailConfigTestResponse"
      },
      "EmailProviderType": {
        "type": "string",
        "enum": ["sendgrid", "ses", "smtp"],
        "title": "EmailProviderType"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TenantEmailConfigCreate": {
        "properties": {
          "provider": { "$ref": "#/components/schemas/EmailProviderType" },
          "api_key": {
            "type": "string",
            "minLength": 1,
            "title": "Api Key",
            "description": "Raw API key — encrypted before storage"
          },
          "from_email": {
            "type": "string",
            "format": "email",
            "title": "From Email"
          }
        },
        "type": "object",
        "required": ["provider", "api_key", "from_email"],
        "title": "TenantEmailConfigCreate",
        "description": "Body for POST /tenants/{tenant_id}/email-config"
      },
      "TenantEmailConfigFallbackResponse": {
        "properties": {
          "configured": {
            "type": "boolean",
            "title": "Configured",
            "default": false
          },
          "using_platform_default": {
            "type": "boolean",
            "title": "Using Platform Default",
            "default": true
          },
          "platform_provider": {
            "type": "string",
            "title": "Platform Provider"
          },
          "platform_from_email": {
            "type": "string",
            "title": "Platform From Email"
          }
        },
        "type": "object",
        "required": ["platform_provider", "platform_from_email"],
        "title": "TenantEmailConfigFallbackResponse",
        "description": "Returned when tenant has no custom config."
      },
      "TenantEmailConfigResponse": {
        "properties": {
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "provider": { "type": "string", "title": "Provider" },
          "from_email": { "type": "string", "title": "From Email" },
          "credential_hint": { "type": "string", "title": "Credential Hint" },
          "is_active": { "type": "boolean", "title": "Is Active" }
        },
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "provider",
          "from_email",
          "credential_hint",
          "is_active"
        ],
        "title": "TenantEmailConfigResponse",
        "description": "Response — never includes the real API key."
      },
      "TenantEmailConfigUpdate": {
        "properties": {
          "provider": {
            "anyOf": [
              { "$ref": "#/components/schemas/EmailProviderType" },
              { "type": "null" }
            ]
          },
          "api_key": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Api Key",
            "description": "If omitted, existing key is kept"
          },
          "from_email": {
            "anyOf": [
              { "type": "string", "format": "email" },
              { "type": "null" }
            ],
            "title": "From Email"
          },
          "is_active": {
            "anyOf": [{ "type": "boolean" }, { "type": "null" }],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "TenantEmailConfigUpdate",
        "description": "Body for PUT /tenants/{tenant_id}/email-config"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" },
          "input": { "title": "Input" },
          "ctx": { "type": "object", "title": "Context" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": { "HTTPBearer": { "type": "http", "scheme": "bearer" } }
  }
}